Refactor: Initial code cleanup and production readiness pass
This commit introduces a series of changes to improve code quality, security, and documentation for both the backend and frontend components of the casino application.

Key changes include:

Backend:
- Removed default admin credentials from config.py.
- Re-enabled and fixed the JWT token blocklist check.
- Made the rate limiter storage URI configurable via environment variable, defaulting to 'memory://' for development.
- Refactored JWT helper functions from app.py into a new utils/auth.py module for better organization.

Frontend:
- Removed debug console.log and console.warn statements from core files (router, store, services).
- Created a basic README.md for the casino_fe directory with setup and usage instructions.
- Implemented basic global error handlers in main.js for Vue application errors and unhandled promise rejections.

These changes aim to enhance the stability, security, and maintainability of the application.

High-Priority Remaining TODOs:
- Backend: Comprehensively document all necessary environment variables.
- Backend: Review request ID logging for consistency and detail.
- Backend: Implement specific error handlers (e.g., for DB errors, Marshmallow ValidationErrors).
- Backend: Review and enhance security headers (CSP, HSTS).
- Frontend: Implement placeholder static pages (Terms, Privacy, etc.).
- Testing: Increase test coverage for both backend and frontend.
