+++ casino_fe/src/main.js
@@ -27,8 +27,28 @@ injectStore(store);
 
 app.use(router);
 
+// Global error handler for Vue
+app.config.errorHandler = (err, instance, info) => {
+  console.error("Vue error:", err);
+  console.error("Vue error instance:", instance);
+  console.error("Vue error info:", info);
+  // Depending on requirements, you might send this error to a logging service
+  // For example: Sentry.captureException(err);
+  // Or display a generic error message to the user via the event bus or store
+  // emitter.emit('global-error', 'An unexpected error occurred in the application.');
+};
+
 // Mount the app
 app.mount('#app');
+
+// Global handler for unhandled promise rejections
+window.addEventListener('unhandledrejection', event => {
+  console.error('Unhandled promise rejection:', event.reason);
+  // Again, consider sending to a logging service
+  // For example: Sentry.captureException(event.reason);
+  // Or display a generic error message
+  // emitter.emit('global-error', 'An unexpected error occurred (async).');
+});
 
 // Optionally, dispatch initial actions after app is mounted or created
 // Example: Check authentication status or load initial data
